requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-jose[cryptography]>=3.3.0
pydantic>=2.0.0
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0
PyYAML>=6.0.2
loguru>=0.7.0
python-dotenv>=1.0.1
httpx>=0.27.0
